Move CSR related classes out of resources package to prepare for factoring out a 'core' module for shared infrastructure across feature verticals.
