+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/resources/ResourceSelectorLoadingItem.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2016 The Android Open Source Project
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,7 +27,7 @@ import javax.swing.SwingConstants;
 import org.jetbrains.annotations.NotNull;
 
 /** Displays UI similar to "loading..." when an resource manager call is in progress. */
-class ResourceSelectorLoadingItem extends JPanel {
+public class ResourceSelectorLoadingItem extends JPanel {
 
   private JLabel progressIcon;
