+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/RepositoryRemotePanel.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 Google Inc. All Rights Reserved.
+ * Copyright 2017 Google Inc. All Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.google.cloud.tools.intellij.resources;
+package com.google.cloud.tools.intellij.vcs;
 
 import com.intellij.openapi.util.text.StringUtil;
 import git4idea.repo.GitRepository;
