+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/vcs/SetupCloudRepository.form
@@ -34,13 +34,13 @@
           <text resource-bundle="messages/CloudToolsBundle" key="cloud.repository.dialog.repository.label"/>
         </properties>
       </component>
-      <component id="9f463" class="com.google.cloud.tools.intellij.resources.RepositorySelector" binding="repositorySelector" custom-create="true">
+      <component id="9f463" class="com.google.cloud.tools.intellij.vcs.RepositorySelector" binding="repositorySelector" custom-create="true">
         <constraints>
           <grid row="1" column="1" row-span="1" col-span="3" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
         <properties/>
       </component>
-      <nested-form id="a83e1" form-file="com/google/cloud/tools/intellij/resources/RepositoryRemotePanel.form" binding="remoteNameSelector" custom-create="true">
+      <nested-form id="a83e1" form-file="com/google/cloud/tools/intellij/vcs/RepositoryRemotePanel.form" binding="remoteNameSelector" custom-create="true">
         <constraints>
           <grid row="2" column="1" row-span="1" col-span="1" vsize-policy="3" hsize-policy="3" anchor="0" fill="1" indent="0" use-parent-layout="false"/>
         </constraints>
